added a funtion to fetch created nfts, refractored old codebase and fixed small bugs in it.

+++ contracts/NFTMarketplace.sol
@@ -48,7 +48,7 @@ contract NFTMarketplace is ERC721URIStorage {
         return lp;
     }
 
-    function createToken(string memory tokenURI, uint256 price)public payable returns (uint){
+    function createToken(string memory tokenURI, uint256 price) public payable returns (uint256){
         _tokenIds.increment();
         uint256 newTokenId = _tokenIds.current();
